requirements.txt
django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
django-environ>=0.11
psycopg2-binary>=2.9     # Driver PostgreSQL
pillow>=10.0              # Untuk ImageField (tanda tangan user)
python-magic>=0.4         # Validasi tipe file upload
openpyxl>=3.1             # Export Excel
WeasyPrint>=60            # Generate PDF report (opsional, bisa diganti reportlab)
drf-nested-routers